+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Changed
 - Number of expected recipients is now calculated in the background on the edit sendout screen to prevent slow page load.
+- Settings item is removed from navigation if `allowAdminChanges` is disabled.
 - Improved how the `@web` alias is determined to be in the site or asset URLs in the preflight check.
 - Changed `Craft::warning` to `Craft::info` when logging user actions.